Update all use cases

+++ src/modules/user/useCases/projectUseCase/services/DeleteProjectService.ts
@@ -5,6 +5,12 @@ class DeleteProjectService {
     async execute(id_project: string) {
         const projectsRepository = getCustomRepository(ProjectsRepository);
 
+        const project = await projectsRepository.findOne(id_project);
+
+        if (!project) {
+            throw new Error("Product does not exists!");
+        }
+
         await projectsRepository.delete(id_project);
     }
 }
